Allow the TF Lite Tasks library to optionally use the TF Lite C API instead of the TF Lite C++ API, if built with '--copt=-DTFLITE_USE_C_API'.
As part of that change, encapsulate all the dependencies of the Tasks library
on the C++ API in TfLiteEngine.  In particular add nested typedefs 'Model',
'Interpreter', and 'InterpreterWrapper' inside TfLiteEngine, and use those,
rather than the 'tflite::' versions, in the rest of the Tasks library.

Also add methods for getting the interpreter inputs and outputs to
TfLiteEngine, and use those in the rest of the Tasks library.

Add separate build targets for some parts of the TF Lite C++ API,
to allow the TF Lite Tasks libraries to depend on those, without
dragging in the whole of the C++ API. In particular:
  - error_reporter.{h,cc} (ErrorReporter)
  - stderr_reporter.{h,cc} (StderrReporter)
  - op_resolver.{h,cc} (OpResolver)
  - mutable_op_resolver.{h,cc} (MutableOpResolver)
  - kernels/register.{h,cc} (BuiltinOpResolver)

Also move TfLiteVerifier from model_builder.h into a separate header file
verifier_interface.h, so that the TF Lite Tasks library can use it
without needing to include model_builder.h from the C++ API.

PiperOrigin-RevId: 327810368